create the CRUD with topics with the command scaffold
