+++ app/helpers/api/v1/topics_helper.rb
