+++ app/models/topic.rb
@@ -1,3 +1,17 @@
+# == Schema Information
+#
+# Table name: topics
+#
+#  id         :bigint           not null, primary key
+#  image      :string           not null
+#  name       :string           not null
+#  created_at :datetime         not null
+#  updated_at :datetime         not null
+#
+# Indexes
+#
+#  index_topics_on_name  (name) UNIQUE
+#
 class Topic < ApplicationRecord
   validates :name, presence: true, uniqueness: true
   validates :image, presence: true
