+++ app/views/api/v1/topics/_topic.json.jbuilder
@@ -1,0 +1,2 @@
+json.extract! topic, :id, :name, :image
+json.url api_v1_topic_url(topic, format: :json)
