+++ app/views/api/v1/topics/index.json.jbuilder
@@ -1,0 +1,1 @@
+json.array! @topics, partial: 'api/v1/topics/topic', as: :topic
