+++ app/views/api/v1/topics/show.html.erb
