+++ app/views/api/v1/topics/show.json.jbuilder
@@ -1,0 +1,1 @@
+json.partial! 'api/v1/topics/topic', topic: @topic
