--- app/views/topics/_topic.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! topic, :id, :name, :image, :created_at, :updated_at
-json.url topic_url(topic, format: :json)
