--- app/views/topics/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.array! @topics, partial: "topics/topic", as: :topic
