--- app/views/topics/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! "topics/topic", topic: @topic
