+++ config/routes.rb
@@ -1,9 +1,10 @@
 Rails.application.routes.draw do
-  scope 'api', defaults: { format: :json } do
-    scope 'v1' do
+  namespace :api, defaults: { format: 'json' } do
+    namespace :v1 do
       resources :topics
     end
   end
+
   mount_devise_token_auth_for 'User', at: 'api/v1', controllers: {
     registrations: 'api/v1/registrations'
   }
