+++ db/migrate/20210804004008_add_gender_field_to_users.rb
@@ -1,5 +1,5 @@
 class AddGenderFieldToUsers < ActiveRecord::Migration[6.1]
   def change
-    add_column :users, :gender, :integer, :null => false
+    add_column :users, :gender, :integer, null: false, default: 0
   end
 end
