+++ db/migrate/20210805204645_create_topics.rb
@@ -2,7 +2,7 @@ class CreateTopics < ActiveRecord::Migration[6.1]
   def change
     create_table :topics do |t|
       t.string :name, null: false
-      t.string :image
+      t.string :image, null: false
 
       t.timestamps
     end
